I am moving away from dynamic memory I am going to use simd vector with static allocation it almost works only comparasion functions doesn't work :(

+++ include/config.hpp
@@ -4,7 +4,7 @@
 #include <experimental/simd>
 
 // specific number types used by neurons
-typedef float number;
+typedef long double number;
 
 #define MAX_SIMD_VECTOR_SIZE std::experimental::simd_abi::max_fixed_size<number>
